requirements.txt
fastapi
uvicorn
pandas
scikit-learn
matplotlib
openai
python-multipart
sentence-transformers
faiss-cpu
python-dotenv
